refactor: PostCardをコンテンツ・アクション部分に分割
357行のPostCardを3つのコンポーネントに分割:
- PostCardContent: Markdownプレビュー・画像・コードスニペット表示
- PostCardActions: いいね・コメント・リアクション・ブックマーク操作
- PostCard: ヘッダー（ユーザー情報）+ 上記コンポーネントの組み合わせ（100行に縮小）
